requirements.txt
fyers-apiv3>=2.0.6
pandas>=1.5.3
numpy>=1.24.3
pyyaml>=6.0
pyotp>=2.8.0
schedule>=1.2.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
dash>=2.8.0
plotly>=5.10.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
argparse>=1.4.0
matplotlib>=3.5.0
seaborn>=0.12.0
openpyxl>=3.1.0
